moving service state to enums
