+++ src/zenml/enums.py
@@ -422,3 +422,14 @@ class StackDeploymentProvider(StrEnum):
     AWS = "aws"
     GCP = "gcp"
     AZURE = "azure"
+
+
+class ServiceState(StrEnum):
+    """Possible states for the service and service endpoint."""
+
+    INACTIVE = "inactive"
+    ACTIVE = "active"
+    PENDING_STARTUP = "pending_startup"
+    PENDING_SHUTDOWN = "pending_shutdown"
+    ERROR = "error"
+    SCALED_TO_ZERO = "scaled_to_zero"
