+++ src/zenml/integrations/databricks/services/databricks_deployment.py
@@ -31,6 +31,7 @@ from databricks.sdk.service.serving import (
 from pydantic import Field
 
 from zenml.client import Client
+from zenml.enums import ServiceState
 from zenml.integrations.databricks.flavors.databricks_model_deployer_flavor import (
     DatabricksBaseConfig,
 )
@@ -38,7 +39,7 @@ from zenml.integrations.databricks.utils.databricks_utils import (
     sanitize_labels,
 )
 from zenml.logger import get_logger
-from zenml.services import ServiceState, ServiceStatus, ServiceType
+from zenml.services import ServiceStatus, ServiceType
 from zenml.services.service import BaseDeploymentService, ServiceConfig
 
 logger = get_logger(__name__)
