+++ src/zenml/integrations/seldon/services/seldon_deployment.py
@@ -22,6 +22,7 @@ import requests
 from pydantic import Field, ValidationError
 
 from zenml import __version__
+from zenml.enums import ServiceState
 from zenml.integrations.seldon.seldon_client import (
     SeldonClient,
     SeldonDeployment,
@@ -31,7 +32,7 @@ from zenml.integrations.seldon.seldon_client import (
 )
 from zenml.logger import get_logger
 from zenml.services.service import BaseDeploymentService, ServiceConfig
-from zenml.services.service_status import ServiceState, ServiceStatus
+from zenml.services.service_status import ServiceStatus
 from zenml.services.service_type import ServiceType
 
 logger = get_logger(__name__)
