+++ src/zenml/services/service.py
@@ -34,10 +34,11 @@ from uuid import UUID
 from pydantic import ConfigDict
 
 from zenml.console import console
+from zenml.enums import ServiceState
 from zenml.logger import get_logger
 from zenml.services.service_endpoint import BaseServiceEndpoint
 from zenml.services.service_monitor import HTTPEndpointHealthMonitor
-from zenml.services.service_status import ServiceState, ServiceStatus
+from zenml.services.service_status import ServiceStatus
 from zenml.services.service_type import ServiceType
 from zenml.utils import source_utils
 from zenml.utils.typed_model import BaseTypedModel
