+++ src/zenml/utils/server_utils.py
@@ -13,10 +13,10 @@
 #  permissions and limitations under the License.
 """Utility functions for ZenML servers."""
 
-from typing import Optional
+from typing import Optional, TYPE_CHECKING
 
-from zenml.zen_server.deploy import LocalServerDeployment
-from zenml.zen_server.deploy.exceptions import ServerDeploymentNotFoundError
+if TYPE_CHECKING:
+    from zenml.zen_server.deploy import LocalServerDeployment
 
 
 def get_local_server() -> Optional["LocalServerDeployment"]:
@@ -29,6 +29,7 @@ def get_local_server() -> Optional["LocalServerDeployment"]:
         found.
     """
     from zenml.zen_server.deploy.deployer import LocalServerDeployer
+    from zenml.zen_server.deploy.exceptions import ServerDeploymentNotFoundError
 
     deployer = LocalServerDeployer()
     try:
